Maintenance commit. Test with multiple resolutions

+++ include/lad_core.hpp
@@ -108,7 +108,7 @@ namespace lad
         int computeExclusionMap(std::string raster, std::string kernel, std::string dst);
         int computeMeanSlopeMap(std::string raster, std::string kernel, std::string mask, std::string dst);
 
-        int compareLayer(std::string src, std::string dst, int threshold, int cmpop); // apply scalar threshold to src raster and store resulting raster in dst layer
+        int compareLayer(std::string src, std::string dst, double threshold, int cmpop); // apply scalar threshold to src raster and store resulting raster in dst layer
         int lowpassFilter(std::string src, std::string dst, cv::Size filterSize, int filterType = 0, double nodata = 0); // apply lowpass filter to input raster Layer and stores the resulting raster in dst Layer
         int computeHeight(std::string src, std::string dst, cv::Size filterSize, int filterType = 0);
